Refactor aplication and on close
