+++ src/app/page.tsx
@@ -1,9 +1,9 @@
 "use client";
 
-import Sidebar from "./components/sidebar/sidebar";
-import Booking from "./components/booking";
-import Modal from "./components/common/containers/modal";
-import BoatSvg from "./assets/svgs/BoatSvg";
+import Sidebar from "../components/sidebar/sidebar";
+import Booking from "../components/booking";
+import Modal from "../components/common/containers/modal";
+import BoatSvg from "../assets/svgs/BoatSvg";
 import Script from "next/script";
 
 export default function Home() {
@@ -12,9 +12,7 @@ export default function Home() {
   };
   return (
     <>
-      <Script
-        src="https://gateway.sumup.com/gateway/ecom/card/v2/sdk.js"
-      />
+      <Script src="https://gateway.sumup.com/gateway/ecom/card/v2/sdk.js" />
       <Modal isOpen={true}>
         <div
           className="flex flex-col  gap-8 items-center justify-center w-full h-full bg-white fixed top-0 left-0 z-50 cursor-pointer"
