+++ src/assets/svgs/BoatSvg.tsx
