+++ src/assets/svgs/CloseSvg.tsx
