+++ src/assets/svgs/ImageSvg.tsx
