+++ src/assets/svgs/InfoSvg.tsx
