+++ src/assets/svgs/PlayerSvg.tsx
