+++ src/assets/svgs/SpinerSvg.tsx
