+++ src/components/common/containers/submit-button.tsx
@@ -1,4 +1,4 @@
-import SpinnerSVG from "@/app/assets/svgs/SpinerSvg";
+import SpinnerSVG from "@/assets/svgs/SpinerSvg";
 import Spinner from "./spinner";
 
 export default function SubmitButton({
