+++ src/components/modals/successModal.tsx
@@ -1,7 +1,7 @@
 "use client";
 import Image from "next/image";
-import EmailImage from "@/app/assets/email 1.png";
-import Modal from "@/app/components/common/containers/modal";
+import EmailImage from "@/assets/email 1.png";
+import Modal from "@/components/common/containers/modal";
 
 export default function SuccessModal({
   isOpen,
