+++ src/components/modals/videoModal.tsx
@@ -1,8 +1,8 @@
 "use client";
 
 import React, { useEffect, useState } from "react";
-import Modal from "@/app/components/common/containers/modal";
-import ReloadSvg from "@/app/assets/svgs/ReloadSvg";
+import Modal from "@/components/common/containers/modal";
+import ReloadSvg from "@/assets/svgs/ReloadSvg";
 import dynamic from "next/dynamic";
 
 const ReactPlayer = dynamic(() => import("react-player/lazy"), { ssr: false });
